pm-agent/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic-settings==2.7.1
confluent-kafka==2.8.0
py-eureka-client==0.11.10
openai==1.61.0
supabase==2.12.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-ollama>=0.2.0
langgraph>=0.2.0
sse-starlette>=2.0.0
psycopg2-binary>=2.9.0
